requirements.txt
gradio
groq
PyPDF2
python-dotenv